requirements.txt
pip~=24.2
pillow~=10.4.0
docxtpl~=0.18.0
openpyxl~=3.1.5
Flask~=3.0.3